Use blivet's fstab_spec as fstab device identifier.
This shares the logic used by anaconda for determining which
identifier to use in /etc/fstab.

The previous code did not have the ability to remove or replace
existing fstab entries if they used a device identifier other
than the canonical device node. This preserves the identifier used
in the existing entry, which facilitates managing entries with
varying types of identifier.