Various test fixes
-
Use correct array syntax for an invalid disk

While both before and after it should trigger an error in the role,
and therefore both are valid tests, a valid array with a nonexistent disk
is more likely to be what was intended, and therefore preferable.
(The former variant actually checked for bad syntax, not for handling
of nonexistent disks.)

-
Test that the role fails, not that it sets blivet_output.failed

The latter would be a valid test as well, but a less important one, and
currently it does not work properly.

Do not use ignore_errors: yes on the test block, it makes all assert
pointless. Use a rescue task instead.

-
Create a file in the test fs to actually detect data loss

-
Do not use ignore_errors: yes in a test block

This makes all test asserts pointless.

Instead catch the error using rescue.

Actually verify that data heve not been destroyed.

-
Remove last instance of ignore_errors: yes

and check for errors properly. Verify again that data have not been lost.

Do not use a pool type of partition, seems to be unimplemented.

Create one volume in the pool, the role does not cope with empty pools.

Note that even with those changes the role application fails -
apparently it refuses to create a VG when there is already a filesystem,
even with safe mode off.

-
Turn off safe mode in cleanup

-
Adjust names of tests to match reality.