Verify that disk volumes are mounted by UUID in fstab
Note that disk volumes mean filesystems directly on disks.

Currently the test fails: on first invocation the filesystem gets mounted by
device name and only on the second by UUID. This should be also aught by a real
idempotence test.

XXX this should be part of the existing verification task lists:
there is already a task called
"Verify that the device identifier appears in /etc/fstab",
but it does not check for this condition.