Determine too_large_size dynamically
To test proper error handling the test playbook needs to attempt to create a
pool that is too large for the underlying device. What is "too large" depends
on the device though. We should not assume that the tests will be run against a
VM with disks created by the provisioning script with the expected size, tests
can be executed in any environment, like on a machine with a number of big
physical disks.

tests_lvm_errors has been using a fixed size as the "too large" size, making it
fail in an environment with disks larger than that. Make it determine the size
dynamically instead. The "too large" size is now just one sector more than the
size of the disk that is going to be used for the test.

Reported by @yizhanglinux.