Add some special handling for volumes' LUKS layer.
This probably merits an explanation. In blivet, the backing device
will have a format of type 'luks'. This is the encrypted, or
backing, device. The next layer out is a LUKSDevice, which
represents the (decrypted/open) device-mapper device. Because the
LUKS layer is optional and can be effectively toggled, there are
many occasions on which it is convenient to look past the LUKS
layer directly to the backing device. For this purpose, all of
blivet's StorageDevice classes have a 'raw_device' property. For
unexcrypted leaf devices, the raw device is the same as the actual
device. For encrypted leaf devices, the raw device points to the
backing device.

In other words, adding _raw_device establishes a line between the
raw/backing/encrypted device and the decrypted/mapped/open device
which makes test validation quite a bit easier.