Add a _create method to BlivetDiskVolume.
Relying on _reformat to create the formatting worked because a
DiskDevice always exists. Now that self._device can be an
optional, non-existent, LUKS layer on top of the disk we have to
make the disk volume class behave more like the other volume
classes -- namely, it has to create its format as part of _create
since we will not always call _reformat (eg: when we've set up,
but not yet created, a new LUKS layer on the disk).