Do not set safe_mode globally
It is better to keep safe mode on. Remove global setting of safe_mode
to off in LUKS tests and set it to off in dividually in tests that need it.

Remove explicit setting of safe_mode in tests for safe_mode: it must be on
by default, so we need to verify that we can rely on the default.