Automatically reduce volume size as needed to fit in pool.
This adds a mechanism for trimming requested volume sizes so that
they fit in their pool. The intention is to accommodate small
shortages arising from metadata space usage, eg: creating a lvm
pool on an unpartitioned 10GiB drive and trying to create a single
100GiB volume in that pool. There is no reason a user of this role
should have to learn about LVM's metadata usage.

The maximum reduction of the volume size is 10% of the requested
size.