Add workaround for incosistent udev DB state after RAID creation
For newly created array in test_raid_volume_options, the spare
drive sometimes doesn't have the linux_raid_member FS type in the
udev database which means Blivet doesn't recognize it as a part
of the array. Forcing udev to rescan the devices after creating
the array should fix this issue.