add support for mount_options (#270)
* add support for mount_options

When support for argument validation was added, that support did not
include the `mount_options` parameter.  This fix adds back that
parameter.  In addition, the volume module arguments are refactored
so that the common volume parameters such as `mount_options` can be
specified in one place.

This adds a test for the `mount_options` parameter, and adds
verification for that parameter.

* only checkout mount_options if requested