Improve skip tags handling in the tests_lvm_pool_members tests
In case "--skip-tags tests::nvme" is set when running ansible,
tests_lvm_pool_members_nvme_generated.yml is not "executed",
but the values and the conditions are evaluated. Then, since
ansible_facts are not set by the setup module, the variables
is_fedora, is_rhel?, and blivet_pkg_version are left undefined,
which causes an error in the "when" condition with "'is_fedora'
is undefined".

To workaround it, we should check if at least one of the variables
is defined before evaluating them.