restrict swap size to less than 128GB on EL7
On EL7 the swap size must be less than 128GB, so change the tests
to restrict the size.  This required introducing a `max_size` parameter
to find_unused_disk.

I also took this opportunity to do some ansible-lint 6.x cleanup
for the code I touched.  There is still a huge amount of tests/
code that is not ansible-lint 6.x clean.