feat: Support for creating volumes without a FS
Currently whenever volume is created without fs_type specification, a filesystem of default
type (i.e. xfs) is automatically put on it.
This change allows user to prevent FS creation by explicitly using "unformatted" value as a fs_type option.
In the same manner it also allows to remove an existing FS. To be able to do that the safe mode has to be off.