test: Library checking script for tests
Storage role development often relies on blivet library and changes in
it. Whether or not is specific feature supported by blivet is usually determined
by its version. That is a cumbersome process, especially when the feature
has not yet been added into blivet and the version has to be guessed.

Added script verifies existence of the feature by using python
introspection and asking for existence of specific item in the library
(e.g. 'blivet.formats.lvmpv.LVMPhysicalVolume.grow_to_fill').
The script is supposed to be used for the tests only.