fix: check mode should not give an error
Cause: The blivet module was not looking for the check_mode flag before
reporting an error when the supporting modules could not be loaded.

Consequence: The storage role would give an error in check mode in the
"Get required packages" task.

Fix: See if the role is running in check mode, and do not report an
error if the underlying modules cannot be loaded.  This is expected.

Result: The storage role does not print an error when running in
check mode.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>