fix: there is no kmod-kvdo on EL10 and later
Cause: The blivet module adds `kmod-kvdo` as a required package on
platforms that do not have the pacakge.

Consequence: The package fails to be installed on EL10 and causes
the role to fail with an error.

Fix: Ensure `kmod-kvdo` is not added as a required package on EL10.

Result: The storage role does not give an error when using VDO
on EL10 and later.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>