tests: Add a custom script for getting info about Stratis pools
There are multiple versions of Stratis in systems we support and
it is hard to get all the information to verify test results from
the Stratis cmdline utility in a consistent way. This adds a small
Python script that uses Stratis DBus API to get the information
we need for the tests.