fix: Correctly check PVs for grow_to_fill feature
We need to check the size of the PV format and compare it to the
size of the underlying block device when deciding if the format
size should be grown. The size of the VG (self._device here) is
not relevant for this operation.
This code used to work because of a bug in blivet where size of
the VG was incorrectly calculated for PVs with size not matching
the underlying block device size, this was fixed in blivet (see
https://github.com/storaged-project/blivet/pull/1334) so the
grow_to_fill feature no longer works without this fix.