fix: Fix getting PVs from raid_disks for RAID LVs
Currently the code expects the disks specified in raid_disks are
parent devices of the PVs we are going to use to allocate the RAID
LVs. This might not be true in several cases, for example if the
VG is encrypted (the LUKS device introduces a new layer in the
storage stack) or if partitions are not used (the disk is then the
PV and doesn't have a parent). This fix also allows user to
specify the PV partition instead of the underlying disk.