fix: the encryption_key parameter should not be marked as no_log
Cause: The encryption_key parameter to the blivet module was marked as
no_log.  This causes the output value from the module to be replaced with the
string `VALUE_SPECIFIED_IN_NO_LOG_PARAMETER` which is then used as the value
for the key file in crypttab and in tests.

Consequence: Disk encryption using a key file does not work as the specified
key file is not used.

Fix: The encryption_key parameter does not need to be marked no_log because
it is just the name of the file, not the actual key contents.

Result: Disk encryption works, and the actual key file is used in crypttab
and elsewhere.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>