fix: check for no disks specified and report correct error
Cause: The code did not check for missing or empty disks in the right place.

Consequence: The role would issue a strange error which did not indicate that the
problem is that no disks were specified.

Fix: Check for missing disks early and report appropriate error.

Result: The role will report an appropriate error message if no disks
are specified.

Signed-off-by: Rich Megginson <rmeggins@redhat.com>